refactor: move interfaces from the types folder to their components and delete types folder
