+++ src/components/SearchResults.tsx
@@ -1,5 +1,21 @@
 import { Component } from 'react';
-import { SearchResultsProps, SearchResultsState } from '../types/SearchResultsInterfaces';
+
+export interface SearchResult {
+  name: string;
+  height: string;
+  mass: string;
+  hair_color: string;
+  skin_color: string;
+  eye_color: string;
+  birth_year: string;
+  gender: string;
+}
+
+interface SearchResultsProps {
+  results: SearchResult[];
+}
+
+interface SearchResultsState {}
 
 export default class SearchResults extends Component<SearchResultsProps, SearchResultsState> {
   render() {
