--- src/types/AppInterface.ts
@@ -1,7 +1,0 @@
-import { SearchResult } from './SearchResultsInterfaces';
-
-export interface AppState {
-  results: SearchResult[];
-  loading: boolean;
-  hasError: boolean;
-}
