--- src/types/ErrorBoundaryInterfaces.ts
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export interface Props {
-  children: React.ReactNode;
-}
-
-export interface State {
-  hasError: boolean;
-}
