--- src/types/ErrorButtonInterfaces.ts
@@ -1,7 +1,0 @@
-export interface ThrowErrorButtonProps {
-  onError: () => void;
-}
-
-export interface ThrowErrorButtonState {
-  throwError: boolean;
-}
