--- src/types/SearchInterfaces.ts
@@ -1,7 +1,0 @@
-export interface SearchProps {
-  onSearch: (term: string) => void;
-}
-
-export interface SearchState {
-  searchTerm: string;
-}
